# backend/tic-tac-toe/docs/swagger.yaml
basePath: /
definitions:
...
  dto.ErrorResponse:
    properties:
      message:
        description: Error message.
        example: invalid request body
        type: string
    type: object
...
  dto.JwtRequest:
    properties:
      login:
        description: User login. Must not be empty.
        example: player
        type: string
      password:
        description: User password. Must not be empty.
        example: secret
        type: string
    type: object
  dto.JwtResponse:
    properties:
      accessToken:
        description: Short-lived access token.
        example: eyJhbGciOiJSUzI1NiIsImtpZCI6InRpYy10YWMtdG9lLW1haW4iLCJ0eXAiOiJKV1QifQ...
        type: string
      refreshToken:
        description: Long-lived refresh token.
        example: eyJhbGciOiJSUzI1NiIsImtpZCI6InRpYy10YWMtdG9lLW1haW4iLCJ0eXAiOiJKV1QifQ...
        type: string
      type:
        description: Token type.
        example: Bearer
        type: string
    type: object
...
  dto.RefreshJwtRequest:
    properties:
      refreshToken:
        description: Refresh token.
        example: eyJhbGciOiJSUzI1NiIsImtpZCI6InRpYy10YWMtdG9lLW1haW4iLCJ0eXAiOiJKV1QifQ...
        type: string
    type: object
  dto.SignUpRequest:
    properties:
      login:
        description: User login. Must not be empty.
        example: player
        type: string
      password:
        description: User password. Must not be empty.
        example: secret
        type: string
    type: object
  dto.SignUpResponse:
    properties:
      success:
        description: Registration result.
        example: true
        type: boolean
    type: object
...
info:
  contact: {}
  description: HTTP API for user authentication and tic-tac-toe games.
  title: Tic-Tac-Toe API
  version: "1.0"
paths:
  /auth/sessions:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshJwtRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Logged out from all sessions
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: Request body must be application/json
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Logout failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Logout from all sessions
      tags:
      - auth
    post:
      consumes:
      - application/json
      description: Send login and password as JSON and receive Bearer access and refresh
        tokens.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.JwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/dto.JwtResponse'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid or missing credentials
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: Request body must be application/json
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Authentication failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Authenticate by login and password
      tags:
      - auth
  /auth/sessions/current:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshJwtRequest'
      produces:
      - application/json
      responses:
        "204":
          description: Logged out
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: Request body must be application/json
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Logout failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Logout
      tags:
      - auth
  /auth/tokens/access:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshJwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/dto.JwtResponse'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: Request body must be application/json
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Refresh failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh access token
      tags:
      - auth
  /auth/tokens/refresh:
    post:
      consumes:
      - application/json
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshJwtRequest'
      produces:
      - application/json
      responses:
        "200":
          description: JWT tokens
          schema:
            $ref: '#/definitions/dto.JwtResponse'
        "400":
          description: Invalid JSON
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Invalid refresh token
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "415":
          description: Request body must be application/json
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Refresh failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh refresh token
      tags:
      - auth
...
  /users:
    post:
      consumes:
      - application/json
      description: Creates a user by login and password. Login is trimmed, password
        is stored as a hash.
      parameters:
      - description: Request body
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/dto.SignUpRequest'
      produces:
      - application/json
      responses:
        "201":
          description: User registered
          schema:
            $ref: '#/definitions/dto.SignUpResponse'
        "400":
          description: Invalid JSON or empty login/password
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Login already exists
          schema:
            $ref: '#/definitions/dto.SignUpResponse'
        "415":
          description: Request body must be application/json
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Registration failed
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register a new user
      tags:
      - auth
...
schemes:
- http
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"